# app/account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/account/login": {
            "post": {
                "description": "login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "登陆账号",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "{\"code\":\"0000\",\"message\":\"ok\",\"data\":{\"token\":\"xxx.xx.xx\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "2101": {
                        "description": "{\"code\":\"2101\",\"message\":\"name or password error\",\"data\":null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/account/regist": {
            "post": {
                "description": "regist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "regist"
                ],
                "summary": "regist",
                "parameters": [
                    {
                        "description": "登陆账号",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "{\"code\":\"0000\",\"message\":\"ok\",\"data\":{\"token\":\"xxx.xx.xx\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "2101": {
                        "description": "{\"code\":\"2101\",\"message\":\"name or password error\",\"data\":null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/file/upload": {
            "post": {
                "description": "上传文件",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "{\"code\": 0,\"msg\": \"ok\",\"data\":{\"src\": \"http://xx.xxx.com/xxx/df13e16abdd0c8317966dbe06cb20778\"}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "100": {
                        "description": "{\"code\": -100,\"msg\": \"参数错误\",\"data\": null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}